refactor: create Rocket.Chat credentials bean
