+++ src/main/java/de/caritas/cob/userservice/api/adapters/rocketchat/RocketChatClient.java
@@ -1,10 +1,5 @@
 package de.caritas.cob.userservice.api.adapters.rocketchat;
 
-import static java.util.Objects.isNull;
-import static java.util.Objects.nonNull;
-
-import de.caritas.cob.userservice.api.exception.httpresponses.BadRequestException;
-import javax.servlet.http.HttpServletRequest;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.http.HttpEntity;
@@ -13,7 +8,6 @@ import org.springframework.http.HttpMethod;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
-import org.springframework.web.util.WebUtils;
 
 @Component
 @Slf4j
@@ -24,15 +18,15 @@ public class RocketChatClient {
 
   private final RestTemplate restTemplate;
 
-  private final HttpServletRequest httpServletRequest;
+  private final RocketChatCredentials rocketChatCredentials;
 
   private final RocketChatCredentialsProvider rcCredentialHelper;
 
   public RocketChatClient(@Qualifier("rocketChatRestTemplate") final RestTemplate restTemplate,
-      final HttpServletRequest httpServletRequest,
+      final RocketChatCredentials rocketChatCredentials,
       final RocketChatCredentialsProvider rocketChatCredentialsProvider) {
     this.restTemplate = restTemplate;
-    this.httpServletRequest = httpServletRequest;
+    this.rocketChatCredentials = rocketChatCredentials;
     this.rcCredentialHelper = rocketChatCredentialsProvider;
   }
 
@@ -73,23 +67,11 @@ public class RocketChatClient {
 
   private HttpHeaders httpHeaders(String userId) {
     var httpHeaders = new HttpHeaders();
-    httpHeaders.add(HEADER_AUTH_TOKEN, rcToken());
+    httpHeaders.add(HEADER_AUTH_TOKEN, rocketChatCredentials.getRocketChatToken());
     httpHeaders.add(HEADER_USER_ID, userId);
 
     return httpHeaders;
   }
 
-  private String rcToken() {
-    var rcToken = httpServletRequest.getHeader("rcToken");
-    if (isNull(rcToken)) {
-      var cookie = WebUtils.getCookie(httpServletRequest, "rc_token");
-      if (nonNull(cookie)) {
-        rcToken = cookie.getValue();
-      } else {
-        throw new BadRequestException("Neither rcToken in header nor rc_token in cookie.");
-      }
-    }
 
-    return rcToken;
-  }
 }
